♻️ refactor work items management in dedicated hook
